+++ CHANGELOG.rst
@@ -3,6 +3,22 @@ Changelog
 ================================
 
 
+Version 0.2.0 (29-09-2019)
+===========================================================
+
+Added
+````````````````````````````````
+
+* Added proxy support
+* Change wallpaper support
+
+Changed
+````````````````````````````````
+
+* Moved to Google Chrome
+* Fixed memory issues and browser other problems.
+
+
 Version 0.1.0 (15-03-2019)
 ===========================================================